Better handle the tax address under edge cases
The tax address is sometimes not known. This means we need to
make a bit better use of the safe navigation operator.

Also Solidus has made the tax address a distinct object from a
normal spree address in the event we are using the default address
(i.e. before a user has a known address).

While I sort of get what they are doing the tax lookup really needs
the full address object. Since Solidus still returns a normal address
object once it is known it doesn't seem like it would hurt anything for
the default tax address to also be a full address object.